Refactor calendar logic into reusable hooks and utilities
Moved calendar-related logic from `WeeklyCalendar` into a custom hook (`useCalendar`) and shared utility files for better modularity. This simplifies the component structure, centralizes logic, and improves code reuse across the application.
